+++ src/routes/(admin)/admin/login/+page.server.ts
@@ -1,48 +1,54 @@
-import { redirect } from '@sveltejs/kit';
+import { fail, redirect } from '@sveltejs/kit';
 import type { PageServerLoad, Actions } from './$types';
+import { GenerateAdminJWE, DecryptAdminJWE } from '$lib/server/auth/jwt';
+import { VerifyAdminCredentials } from '$lib/server/auth/verify';
 
-export const load: PageServerLoad = ({ cookies }) => {
+export const load: PageServerLoad = async ({ cookies }) => {
 
-    console.log("PAGE LOAD CHECK FOR AUTH")
-
-    // check for cookie
     const authCookie = cookies.get("auth");
 
-    console.log("authCookie", authCookie);
+    if (!authCookie) {
+        return
+    }
 
-    // validate jwt in cookie
+    const claims = await DecryptAdminJWE(authCookie);
+    if (claims) {
+        // didn't pass decryption checks, do nothing
+        if (VerifyAdminCredentials(claims.password, claims.pin)) {
+            // if credentials match then redirect to /admin
+            redirect(307, '/admin');
+        }
+    }
+    cookies.delete("auth", { path: "/admin" })
 
-    // if authorized redirect to /admin
-
-    return {
-
-    };
+    return
 };
 
 export const actions = {
     default: async ({ cookies, request }) => {
 
         const data = await request.formData();
-        const pin = data.get("pin");
-        const password = data.get("password");
-
-        console.log("pin", pin);
-        console.log("pass", password);
+        const pin = data.get("pin").toString();
+        const password = data.get("password").toString();
 
         // check login credentials
+        if (VerifyAdminCredentials(password, pin)) {
+            // generate jwt
+            const jwe = await GenerateAdminJWE({ password, pin }, "2days");
 
-        // generate jwt
+            // create cookie with jwe
+            cookies.set("auth", jwe, {
+                httpOnly: true,
+                maxAge: 60 * 60 * 24 * 5,
+                // secure: true,
+                path: "/"
+            });
 
-        // create cookie with jwt
-        cookies.set("auth", "asjkdbfjkasdbfjkasdf", {
-            httpOnly: true,
-            maxAge: 60 * 60 * 2,
-            // secure: true,
-            path: "/admin"
-        })
+            // if credentials match then redirect to /admin
+            redirect(307, '/admin');
+        }
 
-        // redirect to /admin
-        // redirect(307, '/admin');
+        return fail(400, { error: "Invalid Credentials" });
 
     },
 } satisfies Actions;