+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2023-01-01
+
+### Added
+
+- a lot more work on the admin routes
+- started the foundation for jwt and cookie auth
+- started adding some api routes
+- started working on the server side handle hook to place auth on api routes
+
 ## [0.1.2] - 2023-12-24
 
 ### Added